+++ app/src/main/java/com/udacity/asteroidradar/database/DatabaseEntities.kt
@@ -1,0 +1,20 @@
+package com.udacity.asteroidradar.database
+
+import android.os.Parcelable
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+import kotlinx.parcelize.Parcelize
+
+@Entity(tableName = "asteroid_table")
+data class AsteroidEntity(
+    @PrimaryKey
+    val id: Long,
+    val codename: String,
+    val closeApproachDate: String,
+    val absoluteMagnitude: Double,
+    val estimatedDiameter: Double,
+    val relativeVelocity: Double,
+    val distanceFromEarth: Double,
+    val isPotentiallyHazardous: Boolean)
+
+
