+++ app/src/main/java/com/udacity/asteroidradar/database/RoomDatabase.kt
@@ -7,19 +7,6 @@ import androidx.room.*
 import kotlinx.parcelize.Parcelize
 
 
-@Parcelize
-@Entity(tableName = "asteroid_table")
-data class AsteroidEntity(
-                        @PrimaryKey
-                          val id: Long,
-                          val codename: String,
-                          val closeApproachDate: String,
-                          val absoluteMagnitude: Double,
-                          val estimatedDiameter: Double,
-                          val relativeVelocity: Double,
-                          val distanceFromEarth: Double,
-                          val isPotentiallyHazardous: Boolean) : Parcelable
-
 @Dao
 interface AsteroidDao{
 @Query("Select * from asteroid_table Order by closeApproachDate")
